RGS Motor 13 Demo && Editing main.c
I will be holding off on the LCD Display for now and will program for the RGS13 Motor first.

NOTE: That the motor demo here works for the ATmega328p and not the ATmega328pb currently with the difference being that the 328pb has a higher frequency (20 MHz vs 16 MHz). This can be adjusted for the 328pb fairly easy.

So far me and Eric managed to get the 8-bit timer to work for now, however we plan on getting it to work with a 16-bit timer since we can get more accuracy with a 16-bit timer. Unlike the ATmega328pb, the ATmega328p has the disadvantage of only having a single 16-bit timer meanwhile the ATmega328pb has 3x 16-bit timers.